tests/wpt/tests/css/css-text/parsing/text-wrap-style-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text: getComputedStyle().textWrapStyle</title>
<link rel="author" title="Tim Nguyen" href="https://github.com/nt1m">
<link rel="help" href="https://drafts.csswg.org/css-text-4/#text-wrap-style">
<meta name="assert" content="text-wrap-style computed value is auto | balance | pretty | stable">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("text-wrap-style", "auto");
test_computed_value("text-wrap-style", "balance");
test_computed_value("text-wrap-style", "stable");
</script>
</body>
</html>
