third_party/blink/web_tests/external/wpt/css/css-text/parsing/text-wrap-style-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property text-wrap-style value 'auto'
  assert_true: text-wrap-style doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-wrap-style value 'balance'
  assert_true: text-wrap-style doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-wrap-style value 'pretty'
  assert_true: text-wrap-style doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-wrap-style value 'stable'
  assert_true: text-wrap-style doesn't seem to be supported in the computed style expected true got false
Harness: the test ran to completion.
